fix strlen prototype

+++ ft_strlen.c
@@ -6,11 +6,11 @@
 /*   By: dieperei <dieperei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/31 18:30:15 by dieperei          #+#    #+#             */
-/*   Updated: 2022/05/31 18:30:43 by dieperei         ###   ########.fr       */
+/*   Updated: 2022/05/31 18:35:29 by dieperei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	i;
